requirements.txt
fastapi
uvicorn[standard]
tensorflow==2.16.1  
Pillow
numpy
python-multipart